// src/enums/statusEnum.ts
export enum StatusCode {
  OK = 200,
  Created = 201,
  Unauthorized = 401,
  Forbidden = 403,
  BadRequest = 400,
  TooManyRequests = 429,
  Conflict = 409,
  Gone = 410,
  LengthRequired = 411,
  PreconditionFailed = 412,
  NotFound = 404,
  InternalServerError = 500,
  BadGateway = 502,
  ServiceUnavailable = 503,
  GatewayTimeout = 504,
  UnprocessableEntity = 422,
}